update home version
